+++ app/src/main/java/com/durand/dogedex/api/dto/ConsultarMascotaDTO.kt
@@ -5,5 +5,5 @@ import com.squareup.moshi.Json
 
 data class ConsultarMascotaDTO(
     @field:Json(name = "idUsuario")
-    val idUsuario: Long
+    val idUsuario: Int
 )
